+++ simulation/simple_sensor_simulator/src/sensor_simulation/detection_sensor/detection_sensor.cpp
@@ -61,6 +61,8 @@ void DetectionSensor::update(
             object.semantic.type = object.semantic.PEDESTRIAN;
             object.semantic.confidence = 1;
             break;
+          case openscenario_msgs::EntityType::MISC_OBJECT:
+            break;
           default:
             throw SimulationRuntimeError("unsupported entity type!");
             break;