+++ simulation/simple_sensor_simulator/src/simple_sensor_simulator.cpp
@@ -43,6 +43,9 @@ ScenarioSimulator::ScenarioSimulator(const rclcpp::NodeOptions & options)
       &ScenarioSimulator::spawnPedestrianEntity, this, std::placeholders::_1,
       std::placeholders::_2),
     std::bind(
+      &ScenarioSimulator::spawnMiscObjectEntity, this, std::placeholders::_1,
+      std::placeholders::_2),
+    std::bind(
       &ScenarioSimulator::despawnEntity, this, std::placeholders::_1, std::placeholders::_2),
     std::bind(
       &ScenarioSimulator::updateEntityStatus, this, std::placeholders::_1, std::placeholders::_2),
@@ -129,6 +132,16 @@ void ScenarioSimulator::spawnPedestrianEntity(
   res.mutable_result()->set_description("");
 }
 
+void ScenarioSimulator::spawnMiscObjectEntity(
+  const simulation_api_schema::SpawnMiscObjectEntityRequest & req,
+  simulation_api_schema::SpawnMiscObjectEntityResponse & res)
+{
+  misc_objects_.emplace_back(req.parameters());
+  res = simulation_api_schema::SpawnMiscObjectEntityResponse();
+  res.mutable_result()->set_success(true);
+  res.mutable_result()->set_description("");
+}
+
 void ScenarioSimulator::despawnEntity(
   const simulation_api_schema::DespawnEntityRequest & req,
   simulation_api_schema::DespawnEntityResponse & res)
@@ -153,6 +166,15 @@ void ScenarioSimulator::despawnEntity(
     }
   }
   pedestrians_ = pedestrians;
+  std::vector<openscenario_msgs::MiscObjectParameters> misc_objects;
+  for (const auto misc_object : misc_objects_) {
+    if (misc_object.name() != req.name()) {
+      misc_objects.emplace_back(misc_object);
+    } else {
+      found = true;
+    }
+  }
+  misc_objects_ = misc_objects;
   if (found) {
     res.mutable_result()->set_success(true);
   } else {