+++ simulation/traffic_simulator/src/entity/misc_object_entity.cpp
@@ -1,0 +1,61 @@
+// Copyright 2015-2020 Tier IV, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <traffic_simulator/entity/misc_object_entity.hpp>
+
+namespace traffic_simulator
+{
+namespace entity
+{
+MiscObjectEntity::MiscObjectEntity(
+  const std::string & name, const openscenario_msgs::msg::EntityStatus & initial_state,
+  const openscenario_msgs::msg::MiscObjectParameters & params)
+: EntityBase(params.misc_object_category, name, initial_state), params_(params)
+{
+  entity_type_.type = openscenario_msgs::msg::EntityType::MISC_OBJECT;
+}
+
+MiscObjectEntity::MiscObjectEntity(
+  const std::string & name, const openscenario_msgs::msg::MiscObjectParameters & params)
+: EntityBase(params.misc_object_category, name), params_(params)
+{
+  entity_type_.type = openscenario_msgs::msg::EntityType::MISC_OBJECT;
+}
+
+void MiscObjectEntity::onUpdate(double, double)
+{
+  if (status_) {
+    status_->action_status.accel = geometry_msgs::msg::Accel();
+    status_->action_status.twist = geometry_msgs::msg::Twist();
+    status_->action_status.current_action = "static";
+    status_before_update_ = status_;
+  } else {
+    status_before_update_ = status_;
+  }
+}
+
+auto MiscObjectEntity::getBoundingBox() const -> const openscenario_msgs::msg::BoundingBox
+{
+  return params_.bounding_box;
+}
+
+auto MiscObjectEntity::getCurrentAction() const -> const std::string
+{
+  if (status_) {
+    return status_->action_status.current_action;
+  }
+  return "";
+}
+}  // namespace entity
+}  // namespace traffic_simulator